requirements.txt

# Core ML
torch>=2.0.0
torch-geometric>=2.3.0

# Transformers (pinned version required for RadGraph)
transformers==4.57.3
huggingface-hub>=0.34.0,<1.0

# Medical NLP
radgraph>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
protobuf>=3.20.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter (optional, for notebooks)
jupyter
ipykernel
notebook
